requirements.txt
pandas
python-docx